requirement.txt
flask
flask-sqlalchemy
flask-cors
pandas
numpy
scikit-learn
anthropic
python-dotenv
groq
google-genai
reportlab
gunicorn
 